+++ crates/buffer/src/piece_tree/utf8/lines/eol.rs
@@ -40,7 +40,7 @@ impl EndOfLine {
 
     pub fn is_slice_eol(slice: &PieceTreeSlice) -> bool {
         const MAX_EOL_LEN: usize = 3;
-        if slice.len() > MAX_EOL_LEN {
+        if slice.len() > MAX_EOL_LEN as u64 {
             return false;
         }
 
@@ -59,9 +59,9 @@ impl EndOfLine {
         self.as_ref()
     }
 
-    pub fn len(&self) -> usize {
+    pub fn len(&self) -> u64 {
         let string: &str = self.as_ref();
-        string.len()
+        string.len() as u64
     }
 }
