+++ crates/editor/src/common/window.rs
@@ -4,7 +4,7 @@ use crate::editor::windows::Window;
 
 /// Return buffer position at a view point. If point is past line end, the line end position is
 /// returned instead.
-pub(crate) fn pos_at_point(win: &Window, point: Point) -> Option<usize> {
+pub(crate) fn pos_at_point(win: &Window, point: Point) -> Option<u64> {
     if let Some(pos) = win.view().pos_at_point(point) {
         return Some(pos);
     }
